mm-video: vidc: Call free_ion_memory while freeing output buffer
There is a memory leak in the video editor use case where
the destructor is never called. This is due to one of the
changes earlier made for CR-356348. This commit reverts
that part of the change.

CRs-Fixed: 356348
(cherry picked from commit b173c4de0e1dc1c743097b5b08879dcfe0293a39)

Change-Id: Id88f47a75b343b964ee4623b80c23040f2a96b75